fix: お気に入りアーティスト関連のAPIエンドポイントを修正
