+++ src/components/music/ArtistFavoriteButton/ArtistFavoriteButton.tsx
@@ -6,7 +6,7 @@ import styles from "./ArtistFavoriteButton.module.css";
 const ArtistFavoriteButton = ({ id }: { id: number }) => {
   const postFavorite = async () => {
     try {
-      const response = await fetch("/api/favoriteArtists", {
+      const response = await fetch("/api/favorite/artists", {
         method: "POST",
         headers: {
           "Content-Type": "application/json",
